requirements.txt
python-dotenv
fastapi
llama-index
llama-index-llms-together
airportsdata